pyproject.toml
[project]
name = "fastapi-sqlalchemy-postgresql-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.3",
    "bcrypt==4.2.0",
    "dishka==1.3.0",
    "fastapi==0.115.0",
    "loguru==0.7.2",
    "psycopg2-binary==2.9.9",
    "pydantic-settings==2.5.2",
    "pydantic[email]==2.9.2",
    "pyjwt==2.9.0",
    "python-multipart==0.0.12",
    "redis[hiredis]==5.1.1",
    "sqlalchemy[asyncio]==2.0.35",
    "taskiq==0.11.7",
    "taskiq-redis==1.0.2",
    "uvicorn[standard]==0.31.0",
    "websockets==13.1",
    "pytest==8.3.3",
]
